Update lib/environment
Change the imports from prior imports to new imports.
Add a check in _getOrDefault to see if the defaultValue is null or undefined.
Allow JSON decoding of the value if it is of type object.
Update hasDockerEnv and isDocker to check if we are not on linux.
Update hasDockerEnv to use existsSynx instead of statSync.
Add istanbul comments to the environment props.

Add new settings:
1. defaultEnvironmentType - The default environment used by the Roblox URL Constructor, it defaults to Production.
2. defaultProductionBaseUrl - The default URL to use when our environment is Production.
2. defaultStagingBaseUrl - The default URL to use when our environment is Staging.
3. defaultIntegrationBaseUrl - The default URL to use when our environment is Integration.
4. defaultDevelopmentBaseUrl - The default URL to use when our environment is Development.